Moved all header files to include and edited CMakeLists.txt to include the header files and their sources
