+++ hardware_layer/hardware_arduino/src/CMakeLists.txt
@@ -9,7 +9,8 @@ set(CMAKE_CXX_FLAGS "-std=gnu++11 ${CMAKE_CXX_FLAGS}")
 add_definitions(-DUSB_CON)
 
 generate_arduino_firmware(arduino_node
-  SRCS arduino_node.cpp ${ROS_LIB_DIR}/time.cpp
+  SRCS arduino_node.cpp ${ROS_LIB_DIR}/time.cpp /MS5837.cpp /arduino_thrust.cpp
+  HDRS ../include/MS5837.h ..include/arduino_thrust.h
   BOARD uno
   PORT /dev/arduino
 )
