+++ hardware_layer/hardware_arduino/src/Processing.cpp
@@ -1,7 +1,7 @@
 #include <ros/ros.h>
 #include <std_msgs/Int32.h>
 #include <math.h>
-#include "thruster.h"
+#include "../include/thruster.h"
 
 //instances of the class thruster defined in header file
 thruster TEAST,TWEST,TNORTHUP,TSOUTHUP,TNORTHSWAY,TSOUTHSWAY;
